+++ src/lib/firebase.ts
@@ -1,8 +1,11 @@
 import { initializeApp, cert } from 'firebase-admin/app';
 import { getAuth } from 'firebase-admin/auth';
+import path from 'path';
 
 initializeApp({
-  credential: cert('../../my-collection-project-service-account.json'),
+  credential: cert(
+    path.resolve(__dirname, '../../my-collection-project-service-account.json'),
+  ),
 });
 
 export const auth = getAuth();
