+++ src/lib/zod.ts
@@ -1,0 +1,6 @@
+import { extendZodWithOpenApi } from '@asteasolutions/zod-to-openapi';
+import z from 'zod';
+
+extendZodWithOpenApi(z);
+
+export default z;
