--- src/routes.ts
@@ -1,3 +1,0 @@
-import { usersRouter } from './users/router';
-
-export { usersRouter };
