+++ src/routes/auth/signin/index.ts
@@ -1,0 +1,1 @@
+export * from './post';
