+++ src/routes/users/[publicId]/index.ts
@@ -1,0 +1,1 @@
+export * from './get';
