+++ src/routes/users/__root__/index.ts
@@ -1,0 +1,1 @@
+export * from './get';
