+++ src/routes/users/me/get.ts
@@ -1,0 +1,1 @@
+export const hoge = null;
