+++ src/swagger/openapi-schema.ts
@@ -1,3 +1,8 @@
-import userEndpoint from './endpoints/userEndpoint';
+import { generateSchema } from '../utils';
+import { usersRouterHandlers } from '../routes/users';
+import { authRouterHandlers } from '../routes/auth';
 
-export const openAPISchemas = [...userEndpoint];
+export const openAPISchemas = [
+  ...generateSchema('/auth', ['Auth'], authRouterHandlers),
+  ...generateSchema('/users', ['Users'], usersRouterHandlers),
+];
