+++ src/types/index.ts
@@ -1,0 +1,16 @@
+export * from './storage';
+export * from './swagger';
+
+declare global {
+  namespace Express {
+    interface Request {
+      decoded?: {
+        role: string;
+        firebaseUid: string;
+      };
+      validatedBody?: { [key: string]: unknown };
+      validatedParams?: { [key: string]: unknown };
+      validatedQuery?: { [key: string]: unknown };
+    }
+  }
+}
